extract slice-atom-array
