+++ src/index.tsx
@@ -2,3 +2,4 @@ export { atom } from './atom'
 export * from './react-utils'
 export { default as focusAtom } from './focus-atom'
 export { default as localStorageAtom } from './localstorage-atom'
+export * from './slice-atom-array'
