+++ src/react-utils.ts
@@ -2,12 +2,12 @@ import equal from 'fast-deep-equal'
 import React from 'react'
 import { unstable_batchedUpdates } from 'react-dom'
 import { atom } from './atom'
+import { sliceAtomArray } from './slice-atom-array'
 import {
   Atom,
   PrimitiveAtom,
   PrimitiveRemovableAtom,
   ReadableAtom,
-  SetState,
 } from './types'
 
 export function useAtom<Value, Updater>(
@@ -81,57 +81,4 @@ export const useAtomSlice = <T>(
   return keptIndexes
     ? sliced.filter((_, index) => keptIndexes.includes(index))
     : sliced
-}
-
-export const sliceAtomArray = <Value>(
-  atomOfArray: PrimitiveAtom<Array<Value>>,
-): Array<PrimitiveRemovableAtom<Value>> => {
-  const sliceIsRemoved = (index: number) =>
-    index >= atomOfArray.getValue().length
-
-  const getAtomAtIndex = (index: number) => {
-    let cachedValue: Value
-    const newAtom = atom(
-      get => {
-        /** This conceptually signals that the slice at
-         * the index has "completed", and it should not update,
-         * nor return anything other than its cached value.
-         */
-        const newValue = get(atomOfArray)[index]
-        if (sliceIsRemoved(index)) {
-          return cachedValue
-        }
-        cachedValue = newValue
-        return cachedValue
-      },
-      (update: SetState<Value>) => {
-        if (!sliceIsRemoved(index)) {
-          const oldValue = atomOfArray.getValue()[index]
-          const newValue =
-            update instanceof Function ? update(oldValue) : update
-          atomOfArray.update(oldArr => [
-            ...oldArr.slice(0, index),
-            newValue,
-            ...oldArr.slice(index + 1),
-          ])
-        }
-      },
-    )
-    return {
-      ...newAtom,
-      remove: () => {
-        atomOfArray.update(oldArr => [
-          ...oldArr.slice(0, index),
-          ...oldArr.slice(index + 1),
-        ])
-      },
-    }
-  }
-  const getArrayAtLength = (length: number) => {
-    const emptyArray = Array.from(new Array(length))
-    const newValue = emptyArray.map((_, index) => getAtomAtIndex(index))
-    return newValue
-  }
-  const length = atomOfArray.getValue().length
-  return getArrayAtLength(length)
 }
