+++ src/slice-atom-array.ts
@@ -1,0 +1,55 @@
+import { atom } from './atom'
+import { PrimitiveAtom, PrimitiveRemovableAtom, SetState } from './types'
+
+export const sliceAtomArray = <Value>(
+  atomOfArray: PrimitiveAtom<Array<Value>>,
+): Array<PrimitiveRemovableAtom<Value>> => {
+  const sliceIsRemoved = (index: number) =>
+    index >= atomOfArray.getValue().length
+
+  const getAtomAtIndex = (index: number) => {
+    let cachedValue: Value
+    const newAtom = atom(
+      get => {
+        /** This conceptually signals that the slice at
+         * the index has "completed", and it should not update,
+         * nor return anything other than its cached value.
+         */
+        const newValue = get(atomOfArray)[index]
+        if (sliceIsRemoved(index)) {
+          return cachedValue
+        }
+        cachedValue = newValue
+        return cachedValue
+      },
+      (update: SetState<Value>) => {
+        if (!sliceIsRemoved(index)) {
+          const oldValue = atomOfArray.getValue()[index]
+          const newValue =
+            update instanceof Function ? update(oldValue) : update
+          atomOfArray.update(oldArr => [
+            ...oldArr.slice(0, index),
+            newValue,
+            ...oldArr.slice(index + 1),
+          ])
+        }
+      },
+    )
+    return {
+      ...newAtom,
+      remove: () => {
+        atomOfArray.update(oldArr => [
+          ...oldArr.slice(0, index),
+          ...oldArr.slice(index + 1),
+        ])
+      },
+    }
+  }
+  const getArrayAtLength = (length: number) => {
+    const emptyArray = Array.from(new Array(length))
+    const newValue = emptyArray.map((_, index) => getAtomAtIndex(index))
+    return newValue
+  }
+  const length = atomOfArray.getValue().length
+  return getArrayAtLength(length)
+}
